Split sip model objects into separate classes. Rename variable, fix typo
